Add E2E tests to test VxLAN cable driv
This is a temporary workflow added to test support for VxLAN driver
while its being developed. This workflow will run only on PRs with label
'vxlan'. Once https://github.com/submariner-io/submariner/pull/1171 is
merged, this workflow will be removed and VxLAN cable driver testing
will be moved to e2e.yml to be tested on all patches irrespective of the
label.

Signed-off-by: Janki Chhatbar <jchhatba@redhat.com>